refactor(phase-9): remove RelationType, fix hardcoded edge type, eliminate defensive pragmas
- B1: Delete RelationType enum from models.py and switch
  ExtractedRelation.relation_type to EdgeType (the enum already covers
  every relation type that RelationType duplicated).
- B2: Replace the hardcoded 'cites' literal in PageRank with a
  parameterized query using EdgeType.CITES.value.
- B3: The three "defensive" pragmas in unified_graph.py (add_node /
  add_edge IntegrityError fall-through, update_node broad except) were
  removed in commit 1 with proper test coverage via mock-based
  injection of unfamiliar IntegrityError subclasses and a serialization
  failure path. Re-confirmed here with the test file consolidation.
- B4/B5: Delete tests/unit/services/intelligence/test_coverage_extras.py
  and fold every still-relevant test into the matching primary file
  (test_migrations.py, test_unified_graph.py, test_time_series.py,
  test_models.py). The defensive no-op test
  test_update_node_unexpected_exception is dropped (now covered by
  test_update_node_serialization_error in commit 1).

Tests: 4232 passing. Intelligence storage coverage: 99.50%.
