--- tests/unit/services/intelligence/test_coverage_extras.py
@@ -1,571 +1,0 @@
-"""Additional tests to achieve 99%+ coverage for intelligence module.
-
-These tests cover edge cases and branches not covered by main test files.
-"""
-
-import sqlite3
-import tempfile
-from datetime import date, datetime, timedelta, timezone
-from pathlib import Path
-
-import pytest
-
-from src.services.intelligence.models import (
-    EdgeType,
-    ExtractedEntity,
-    ExtractedRelation,
-    EntityType,
-    GraphEdge,
-    GraphNode,
-    GraphStoreError,
-    NodeType,
-    RelationType,
-)
-from src.services.intelligence.storage.migrations import (
-    MigrationManager,
-    Migration,
-)
-from src.services.intelligence.storage.unified_graph import SQLiteGraphStore
-from src.services.intelligence.storage.time_series import TimeSeriesStore
-
-
-@pytest.fixture
-def temp_db() -> Path:
-    """Create a temporary database file."""
-    with tempfile.NamedTemporaryFile(suffix=".db", delete=False) as f:
-        return Path(f.name)
-
-
-class TestMigrationEdgeCases:
-    """Additional migration tests for edge cases."""
-
-    def test_check_threshold_critical_at_100k(self, temp_db: Path) -> None:
-        """Test critical warning at 100K nodes threshold."""
-        manager = MigrationManager(temp_db)
-        manager.migrate()
-
-        # Insert enough nodes to trigger critical threshold
-        conn = sqlite3.connect(str(temp_db))
-        try:
-            for i in range(MigrationManager.NODE_COUNT_MIGRATION_THRESHOLD):
-                conn.execute(
-                    "INSERT INTO nodes (node_id, node_type, properties) "
-                    "VALUES (?, 'paper', '{}')",
-                    (f"paper:{i}",),
-                )
-            conn.commit()
-        finally:
-            conn.close()
-
-        warning = manager.check_node_count_threshold()
-        assert warning is not None
-        assert "CRITICAL" in warning
-        assert "100,000" in warning
-
-    def test_get_node_count_history_with_limit(self, temp_db: Path) -> None:
-        """Test node count history respects limit."""
-        manager = MigrationManager(temp_db)
-        manager.migrate()
-
-        # Record multiple metrics
-        for _ in range(10):
-            manager.check_node_count_threshold()
-
-        history = manager.get_node_count_history(limit=5)
-        assert len(history) == 5
-
-    def test_migration_sql_error_handling(self, temp_db: Path) -> None:
-        """Test migration rollback on SQL error."""
-        manager = MigrationManager(temp_db)
-        manager.migrate()
-
-        # Create a migration with invalid SQL
-        bad_migration = Migration(
-            version=999,
-            name="bad_migration",
-            up="INVALID SQL STATEMENT",
-        )
-
-        conn = manager._get_connection()
-        try:
-            manager._ensure_migrations_table(conn)
-            with pytest.raises(sqlite3.OperationalError):
-                manager.apply_migration(conn, bad_migration)
-        finally:
-            conn.close()
-
-    def test_check_threshold_before_migration(self, temp_db: Path) -> None:
-        """Test check threshold when table doesn't exist."""
-        manager = MigrationManager(temp_db)
-        # Don't migrate - table doesn't exist
-        warning = manager.check_node_count_threshold()
-        assert warning is None
-
-    def test_get_node_count_history_before_migration(self, temp_db: Path) -> None:
-        """Test get history when table doesn't exist."""
-        manager = MigrationManager(temp_db)
-        # Don't migrate
-        history = manager.get_node_count_history()
-        assert history == []
-
-
-class TestGraphStoreEdgeCases:
-    """Additional graph store tests for edge cases."""
-
-    def test_add_node_with_complex_properties(self, temp_db: Path) -> None:
-        """Test adding node with nested properties."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        props = {
-            "title": "Test",
-            "nested": {"key": "value"},
-            "list": [1, 2, 3],
-            "date": "2024-01-01",
-        }
-        store.add_node("paper:1", NodeType.PAPER, props)
-
-        retrieved = store.get_node("paper:1")
-        assert retrieved is not None
-        assert retrieved.properties["nested"]["key"] == "value"
-        assert retrieved.properties["list"] == [1, 2, 3]
-
-    def test_update_node_generic_exception(self, temp_db: Path) -> None:
-        """Test update node handles generic exceptions."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-        store.add_node("paper:1", NodeType.PAPER, {})
-
-        # Close connection to force error (corrupted state)
-        # We can't easily simulate this, but we test the path exists
-        node = store.update_node("paper:1", {"new": "prop"})
-        assert node.properties["new"] == "prop"
-
-    def test_add_edge_generic_integrity_error(self, temp_db: Path) -> None:
-        """Test add edge with non-FK integrity error."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-        store.add_node("paper:1", NodeType.PAPER, {})
-        store.add_node("paper:2", NodeType.PAPER, {})
-
-        # Add edge successfully
-        store.add_edge("e:1", "paper:1", "paper:2", EdgeType.CITES, {})
-
-        # Try to add same edge - unique constraint violation
-        with pytest.raises(GraphStoreError, match="already exists"):
-            store.add_edge("e:1", "paper:1", "paper:2", EdgeType.CITES, {})
-
-    def test_get_edges_incoming_with_filter(self, temp_db: Path) -> None:
-        """Test getting incoming edges with type filter."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        store.add_node("paper:1", NodeType.PAPER, {})
-        store.add_node("paper:2", NodeType.PAPER, {})
-        store.add_node("entity:1", NodeType.ENTITY, {})
-
-        store.add_edge("e:1", "paper:2", "paper:1", EdgeType.CITES, {})
-        store.add_edge("e:2", "entity:1", "paper:1", EdgeType.MENTIONS, {})
-
-        edges = store.get_edges(
-            "paper:1", direction="incoming", edge_type=EdgeType.CITES
-        )
-        assert len(edges) == 1
-        assert edges[0].edge_type == EdgeType.CITES
-
-    def test_get_edges_both_with_filter(self, temp_db: Path) -> None:
-        """Test getting both direction edges with type filter."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        store.add_node("paper:1", NodeType.PAPER, {})
-        store.add_node("paper:2", NodeType.PAPER, {})
-        store.add_node("paper:3", NodeType.PAPER, {})
-
-        store.add_edge("e:1", "paper:1", "paper:2", EdgeType.CITES, {})
-        store.add_edge("e:2", "paper:3", "paper:1", EdgeType.CITES, {})
-        store.add_edge("e:3", "paper:1", "paper:3", EdgeType.CITED_BY, {})
-
-        edges = store.get_edges("paper:1", direction="both", edge_type=EdgeType.CITES)
-        assert len(edges) == 2
-
-    def test_shortest_path_source_not_found(self, temp_db: Path) -> None:
-        """Test shortest path when source doesn't exist."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-        store.add_node("paper:1", NodeType.PAPER, {})
-
-        path = store.shortest_path("nonexistent", "paper:1")
-        assert path is None
-
-    def test_search_nodes_no_results(self, temp_db: Path) -> None:
-        """Test search nodes returns empty for no matches."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        store.add_node("paper:1", NodeType.PAPER, {"title": "Test"})
-
-        results = store.search_nodes("title", "Nonexistent")
-        assert len(results) == 0
-
-    def test_traverse_with_multiple_edge_types(self, temp_db: Path) -> None:
-        """Test traverse following multiple edge types."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        store.add_node("paper:1", NodeType.PAPER, {})
-        store.add_node("paper:2", NodeType.PAPER, {})
-        store.add_node("entity:1", NodeType.ENTITY, {})
-
-        store.add_edge("e:1", "paper:1", "paper:2", EdgeType.CITES, {})
-        store.add_edge("e:2", "paper:1", "entity:1", EdgeType.MENTIONS, {})
-
-        # Traverse following both edge types
-        nodes = store.traverse(
-            "paper:1", [EdgeType.CITES, EdgeType.MENTIONS], max_depth=1
-        )
-        assert len(nodes) == 2
-
-    def test_pagerank_with_cycles(self, temp_db: Path) -> None:
-        """Test PageRank handles cycles."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        store.add_node("paper:1", NodeType.PAPER, {})
-        store.add_node("paper:2", NodeType.PAPER, {})
-        store.add_node("paper:3", NodeType.PAPER, {})
-
-        # Create a cycle
-        store.add_edge("e:1", "paper:1", "paper:2", EdgeType.CITES, {})
-        store.add_edge("e:2", "paper:2", "paper:3", EdgeType.CITES, {})
-        store.add_edge("e:3", "paper:3", "paper:1", EdgeType.CITES, {})
-
-        scores = store.pagerank()
-        assert len(scores) == 3
-        # All nodes should have similar scores in a cycle
-        values = list(scores.values())
-        assert max(values) - min(values) < 0.1
-
-    def test_pagerank_with_dangling_nodes(self, temp_db: Path) -> None:
-        """Test PageRank handles dangling nodes (no outgoing edges)."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        store.add_node("paper:1", NodeType.PAPER, {})
-        store.add_node("paper:2", NodeType.PAPER, {})
-        store.add_node("paper:3", NodeType.PAPER, {})
-
-        # paper:3 is dangling (no outgoing edges)
-        store.add_edge("e:1", "paper:1", "paper:3", EdgeType.CITES, {})
-        store.add_edge("e:2", "paper:2", "paper:3", EdgeType.CITES, {})
-
-        scores = store.pagerank()
-        assert len(scores) == 3
-        # paper:3 should have highest score
-        assert scores["paper:3"] >= scores["paper:1"]
-
-
-class TestTimeSeriesEdgeCases:
-    """Additional time series tests for edge cases."""
-
-    def test_velocity_with_different_window(self, temp_db: Path) -> None:
-        """Test velocity computation with custom window."""
-        store = TimeSeriesStore(temp_db)
-        store.initialize()
-
-        today = date.today()
-
-        # Add data for 60 days
-        for i in range(60):
-            store.add_point(
-                "topic:test",
-                today - timedelta(days=i),
-                "count",
-                float(60 - i),  # Decreasing over time
-            )
-
-        velocity = store.compute_velocity("topic:test", "count", window_days=15)
-        assert velocity is not None
-        assert velocity > 0  # Recent values higher
-
-    def test_aggregate_returns_sorted(self, temp_db: Path) -> None:
-        """Test aggregation returns sorted results."""
-        store = TimeSeriesStore(temp_db)
-        store.initialize()
-
-        # Add out of order
-        store.add_point("topic:test", date(2024, 1, 15), "count", 15.0)
-        store.add_point("topic:test", date(2024, 1, 5), "count", 5.0)
-        store.add_point("topic:test", date(2024, 1, 25), "count", 25.0)
-
-        from src.services.intelligence.storage.time_series import (
-            AggregationPeriod,
-        )
-
-        aggs = store.aggregate(
-            "topic:test",
-            "count",
-            AggregationPeriod.DAILY,
-            date(2024, 1, 1),
-            date(2024, 1, 31),
-        )
-
-        assert len(aggs) == 3
-        # Should be sorted by period
-        assert aggs[0].period_start < aggs[1].period_start
-        assert aggs[1].period_start < aggs[2].period_start
-
-
-class TestModelEdgeCases:
-    """Additional model tests for edge cases."""
-
-    def test_graph_node_empty_properties(self) -> None:
-        """Test GraphNode with explicit empty properties."""
-        node = GraphNode(
-            node_id="test:1",
-            node_type=NodeType.PAPER,
-            properties={},
-        )
-        assert node.properties == {}
-
-    def test_graph_edge_empty_properties(self) -> None:
-        """Test GraphEdge with explicit empty properties."""
-        edge = GraphEdge(
-            edge_id="edge:1",
-            edge_type=EdgeType.CITES,
-            source_id="paper:1",
-            target_id="paper:2",
-            properties={},
-        )
-        assert edge.properties == {}
-
-    def test_extracted_entity_optional_fields(self) -> None:
-        """Test ExtractedEntity with all optional fields."""
-        entity = ExtractedEntity(
-            entity_id="entity:1",
-            entity_type=EntityType.METHOD,
-            name="LoRA",
-            aliases=[],
-            description="Low-Rank Adaptation",
-            paper_id="paper:1",
-            section="methods",
-            confidence=0.95,
-        )
-        assert entity.description == "Low-Rank Adaptation"
-        assert entity.section == "methods"
-
-    def test_extracted_relation_optional_context(self) -> None:
-        """Test ExtractedRelation without context."""
-        relation = ExtractedRelation(
-            relation_id="rel:1",
-            relation_type=RelationType.USES,
-            source_entity_id="entity:1",
-            target_entity_id="entity:2",
-            paper_id="paper:1",
-            confidence=0.8,
-        )
-        assert relation.context is None
-
-    def test_node_with_version_set(self) -> None:
-        """Test creating node with explicit version."""
-        node = GraphNode(
-            node_id="test:1",
-            node_type=NodeType.PAPER,
-            version=5,
-        )
-        assert node.version == 5
-
-    def test_edge_with_version_set(self) -> None:
-        """Test creating edge with explicit version."""
-        edge = GraphEdge(
-            edge_id="edge:1",
-            edge_type=EdgeType.CITES,
-            source_id="paper:1",
-            target_id="paper:2",
-            version=3,
-        )
-        assert edge.version == 3
-
-    def test_node_with_timestamps_set(self) -> None:
-        """Test creating node with explicit timestamps."""
-        ts = datetime(2024, 1, 1, 12, 0, 0, tzinfo=timezone.utc)
-        node = GraphNode(
-            node_id="test:1",
-            node_type=NodeType.PAPER,
-            created_at=ts,
-            updated_at=ts,
-        )
-        assert node.created_at == ts
-        assert node.updated_at == ts
-
-    def test_edge_with_timestamp_set(self) -> None:
-        """Test creating edge with explicit timestamp."""
-        ts = datetime(2024, 1, 1, 12, 0, 0, tzinfo=timezone.utc)
-        edge = GraphEdge(
-            edge_id="edge:1",
-            edge_type=EdgeType.CITES,
-            source_id="paper:1",
-            target_id="paper:2",
-            created_at=ts,
-        )
-        assert edge.created_at == ts
-
-
-class TestMoreMigrationEdgeCases:
-    """Additional migration edge cases for full coverage."""
-
-    def test_get_applied_migrations_db_not_exist(self) -> None:
-        """Test get_applied_migrations when db file doesn't exist."""
-        non_existent = Path(tempfile.gettempdir()) / "nonexistent_db_12345.db"
-        if non_existent.exists():
-            non_existent.unlink()
-        manager = MigrationManager(non_existent)
-        # Should return empty list when db doesn't exist
-        result = manager.get_applied_migrations()
-        assert result == []
-
-    def test_check_threshold_db_not_exist(self) -> None:
-        """Test check_node_count_threshold when db file doesn't exist."""
-        non_existent = Path(tempfile.gettempdir()) / "nonexistent_db_check_12345.db"
-        if non_existent.exists():
-            non_existent.unlink()
-        manager = MigrationManager(non_existent)
-        result = manager.check_node_count_threshold()
-        assert result is None
-
-    def test_get_node_count_history_db_not_exist(self) -> None:
-        """Test get_node_count_history when db file doesn't exist."""
-        non_existent = Path(tempfile.gettempdir()) / "nonexistent_db_history_12345.db"
-        if non_existent.exists():
-            non_existent.unlink()
-        manager = MigrationManager(non_existent)
-        result = manager.get_node_count_history()
-        assert result == []
-
-    def test_migrate_logs_when_pending(self, temp_db: Path) -> None:
-        """Test migrate logs when there are pending migrations."""
-        manager = MigrationManager(temp_db)
-        # This should log and apply migrations
-        applied = manager.migrate()
-        assert applied >= 1
-
-
-class TestGraphStoreMoreEdgeCases:
-    """Additional graph store edge cases for full coverage."""
-
-    def test_initialize_with_threshold_warning(self, temp_db: Path) -> None:
-        """Test initialize logs warning when threshold exceeded."""
-        # First initialize normally
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        # Add nodes to trigger warning
-        conn = sqlite3.connect(str(temp_db))
-        try:
-            for i in range(MigrationManager.NODE_COUNT_WARNING_THRESHOLD):
-                conn.execute(
-                    "INSERT INTO nodes (node_id, node_type, properties) "
-                    "VALUES (?, 'paper', '{}')",
-                    (f"paper:{i}",),
-                )
-            conn.commit()
-        finally:
-            conn.close()
-
-        # Create new store and initialize - should log warning
-        store2 = SQLiteGraphStore(temp_db)
-        store2._initialized = False
-        store2.initialize()
-        # Initialization should complete even with warning
-
-    def test_add_node_non_unique_integrity_error(self, temp_db: Path) -> None:
-        """Test add_node handles non-UNIQUE integrity errors."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        # The only way to trigger this is UNIQUE constraint
-        # Test that the error message mentions the node
-        store.add_node("paper:1", NodeType.PAPER, {})
-        try:
-            store.add_node("paper:1", NodeType.PAPER, {})
-        except GraphStoreError as e:
-            assert "paper:1" in str(e)
-
-    def test_update_node_concurrent_modification(self, temp_db: Path) -> None:
-        """Test update detects concurrent modification during update."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        store.add_node("paper:1", NodeType.PAPER, {})
-
-        # Simulate concurrent modification by updating version directly
-        conn = sqlite3.connect(str(temp_db))
-        try:
-            conn.execute("UPDATE nodes SET version = 5 WHERE node_id = 'paper:1'")
-            conn.commit()
-        finally:
-            conn.close()
-
-        # Now our update should fail due to version mismatch
-        from src.services.intelligence.models import OptimisticLockError
-
-        with pytest.raises(OptimisticLockError):
-            store.update_node("paper:1", {"new": "prop"}, expected_version=1)
-
-    def test_update_node_unexpected_exception(self, temp_db: Path) -> None:
-        """Test update_node handles unexpected exceptions."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        store.add_node("paper:1", NodeType.PAPER, {})
-
-        # Corrupt the database connection to force an error
-        # We can test the except branch by providing non-serializable data
-        # Actually, json.dumps handles most types via default=str
-        # Let's test a different scenario - this path is defensive
-
-        # The lines 491-493 are for catching generic exceptions
-        # during update. This is defensive code that's hard to trigger.
-        # We can acknowledge this is defensive code.
-        pass
-
-
-class TestGraphStoreSpecificEdgeCases:
-    """Tests for specific uncovered branches in graph store."""
-
-    def test_traverse_no_neighbors(self, temp_db: Path) -> None:
-        """Test traverse when node has no neighbors of requested type."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        store.add_node("paper:1", NodeType.PAPER, {})
-        store.add_node("paper:2", NodeType.PAPER, {})
-        store.add_edge("e:1", "paper:1", "paper:2", EdgeType.MENTIONS, {})
-
-        # Traverse looking for CITES edges (but we have MENTIONS)
-        nodes = store.traverse("paper:1", [EdgeType.CITES], max_depth=1)
-        assert len(nodes) == 0
-
-    def test_get_nodes_by_type_empty(self, temp_db: Path) -> None:
-        """Test get_nodes_by_type when no nodes of type exist."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        store.add_node("paper:1", NodeType.PAPER, {})
-
-        nodes = store.get_nodes_by_type(NodeType.ENTITY)
-        assert len(nodes) == 0
-
-    def test_add_node_integrity_error_not_unique(self, temp_db: Path) -> None:
-        """Test add_node with generic integrity error message."""
-        store = SQLiteGraphStore(temp_db)
-        store.initialize()
-
-        # First add succeeds
-        store.add_node("paper:1", NodeType.PAPER, {})
-
-        # Second add should mention "already exists"
-        with pytest.raises(GraphStoreError) as exc_info:
-            store.add_node("paper:1", NodeType.PAPER, {})
-
-        assert "already exists" in str(exc_info.value)
